use i16 not i32 for conthist
bench: 7347735

+++ src/history.h
@@ -1,6 +1,6 @@
 #pragma once
 class History {
-  int conthist[2][6][64][2][6][64];
+  short int conthist[2][6][64][2][6][64];
   int quiethistory[2][6][64];
   int noisyhistory[2][6][6];
   const int quietlimit = 27000;
